feat(backstage) role curd

+++ sipue-common/common-core/src/main/java/com/sipue/common/core/constants/CommonConstants.java
@@ -64,7 +64,7 @@ public interface CommonConstants {
     /**
      * 目录
      */
-    Integer CATALOG = -1;
+    Long CATALOG = -1L;
 
     /**
      * 菜单
